requirements.txt
selenium
Pillow
pytesseract
webdriver-manager